+++ back-end/src/main/java/com/skynet/skymed/SkymedApplication.java
@@ -1,4 +1,4 @@
-package com.skynet.Skymed;
+package com.skynet.skymed;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
