+++ back-end/src/main/java/com/skynet/skymed/controller/EnderecoController.java
@@ -1,4 +1,4 @@
-package com.skynet.Skymed.controller;
+package com.skynet.skymed.controller;
 
 import java.util.ArrayList;
 
@@ -11,41 +11,40 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.skynet.Skymed.interfaces.IController;
-import com.skynet.Skymed.model.Pessoa;
-import com.skynet.Skymed.repository.PessoaRepository;
- 
+import com.skynet.skymed.interfaces.IController;
+import com.skynet.skymed.model.Endereco;
+import com.skynet.skymed.repository.EnderecoRepository;
 
 @RestController
 
-@RequestMapping("/pessoa")
-public class PessoaController implements IController<Pessoa> {
+@RequestMapping("/endereco")
+public class EnderecoController implements IController<Endereco> {
 
 	@Autowired
-	private PessoaRepository pessoaDB;
+	private EnderecoRepository enderecoDB;
 
 	@GetMapping
 	@Override
-	public ArrayList<Pessoa> getObject() {
-		return (ArrayList<Pessoa>) pessoaDB.findAll();
+	public ArrayList<Endereco> getObject() {
+		return (ArrayList<Endereco>) enderecoDB.findAll();
 	}
 
 	@PostMapping
 	@Override
-	public void postAndPutObject(@RequestBody Pessoa object) {
-		pessoaDB.save(object);
+	public void postAndPutObject(@RequestBody Endereco object) {
+		enderecoDB.save(object);
 	}
 
 	@DeleteMapping("/{id}")
 	@Override
 	public void deleteObject(@PathVariable("id") Integer id) {
-		pessoaDB.deleteById((long)id);
+		enderecoDB.deleteById((long)id);
 	}
 
 	@GetMapping("/{id}")
 	@Override
-	public Pessoa getById(@PathVariable("id") Integer id) {
-		return pessoaDB.getOne((long)id);
+	public Endereco getById(@PathVariable("id") Integer id) {
+		return enderecoDB.getOne((long)id);
 	}
