+++ back-end/src/main/java/com/skynet/skymed/repository/EnderecoRepository.java
@@ -1,8 +1,9 @@
-package com.skynet.Skymed.repository;
+package com.skynet.skymed.repository;
 
 import org.springframework.stereotype.Repository;
 
-import com.skynet.Skymed.model.Endereco;
+import com.skynet.skymed.model.Endereco;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
