+++ back-end/src/main/java/com/skynet/skymed/repository/MedicoRepository.java
@@ -1,8 +1,9 @@
-package com.skynet.Skymed.repository;
+package com.skynet.skymed.repository;
 
 import org.springframework.stereotype.Repository;
 
-import com.skynet.Skymed.model.Medico;
+import com.skynet.skymed.model.Medico;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
