+++ back-end/src/main/java/com/skynet/skymed/repository/PessoaRepository.java
@@ -1,9 +1,9 @@
-package com.skynet.Skymed.repository;
+package com.skynet.skymed.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.skynet.Skymed.model.Pessoa;
+import com.skynet.skymed.model.Pessoa;
 
 @Repository
 public interface PessoaRepository extends JpaRepository<Pessoa, Long> {
